# Add klog.Info/Error support