--- a/internal/e2e/guest/go.mod
+++ b/internal/e2e/guest/go.mod
@@ -3,12 +3,12 @@ module sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e/guest
 go 1.20
 
 require (
-	github.com/wasilibs/nottinygc v0.3.0
+	github.com/wasilibs/nottinygc v0.4.0
+	sigs.k8s.io/kube-scheduler-wasm-extension/guest v0.0.0-00010101000000-000000000000
 	sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto v0.0.0-00010101000000-000000000000
 )
 
 require (
-	github.com/google/go-cmp v0.5.9 // indirect
 	github.com/magefile/mage v1.14.0 // indirect
 	google.golang.org/protobuf v1.30.0 // indirect
 )