--- a/scheduler/go.mod
+++ b/scheduler/go.mod
@@ -40,6 +40,7 @@ require (
 	k8s.io/api v0.27.3
 	k8s.io/apimachinery v0.27.3
 	k8s.io/component-base v0.27.3
+	k8s.io/klog/v2 v2.90.1
 	k8s.io/kubectl v0.27.3
 	k8s.io/kubernetes v1.27.3
 )
@@ -136,7 +137,6 @@ require (
 	k8s.io/controller-manager v0.27.3 // indirect
 	k8s.io/csi-translation-lib v0.0.0 // indirect
 	k8s.io/dynamic-resource-allocation v0.0.0 // indirect
-	k8s.io/klog/v2 v2.90.1 // indirect
 	k8s.io/kms v0.27.3 // indirect
 	k8s.io/kube-openapi v0.0.0-20230501164219-8b0f38b5fd1f // indirect
 	k8s.io/kube-scheduler v0.0.0 // indirect