… examples/advanced/main.wasm
… examples/advanced/plugin/plugin.go
… examples/advanced/plugin/plugin_test.go
… examples/nodenumber/main.go
… examples/nodenumber/main.wasm
… guest/RATIONALE.md
… guest/klog/api/k8s_references_test.go
… guest/klog/api/klog_test.go
… guest/klog/internal/klog_test.go
… guest/klog/klog_test.go
… guest/testdata/all/main.go
… guest/testdata/all/main.wasm
… guest/testdata/filter/main.wasm
… internal/e2e/guest/go.sum
… internal/e2e/guest/guest_test.go
… internal/e2e/scheduler/scheduler_test.go
… scheduler/plugin/host_test.go
… scheduler/plugin/plugin_test.go
… scheduler/plugin/runtime.go
… scheduler/test/testdata.go