#### What type of PR is this?

/kind feature

#### What this PR does / why we need it:

This begins klog support with the simplest api used in plugins: `klog.Info`, `klog.InfoS`, `klog.Error` and `klog.ErrorS`

#### Which issue(s) this PR fixes:

NONE

#### Special notes for your reviewer:

* Note: "severity" is internal in klog, and we don't need to expose it in the API, just use the same constants.
* wasm cannot pass objects to the host. This is why we have to stringify in the guest first. This is most obvious in `klog.InfoS`, where the KVs are encoded NUL-terminated to the host.
* utilities like `klog.KObj` used reflection too much which caused things like crashing the compiler or poor performance. This ports the intent more simply.

#### Does this PR introduce a user-facing change?

NONE

#### What are the benchmark results of this change?

Performance of Simple is wild because it uses the default GC and the plugin is heavy due to protos. Advanced uses [nottinygc](https://github.com/wasilibs/nottinygc) and also separates out code that uses imports so that they can be more easily tested. These notes are in the [READMEs](https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/blob/main/examples/advanced/README.md).

The main thing here is that performance of Advanced/Run is the same when logging is disabled, and it drops 13% (7us) per Run when enabled. This shows the deferred stringification works, but also shows logging has a measurable cost even when only a couple times.

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e/scheduler
                                       │  before.txt  │              now.txt               │
                                       │    sec/op    │    sec/op     vs base              │
Example_NodeNumber/Simple/New-12         63.64m ±  3%   62.56m ±  0%       ~ (p=0.065 n=6)
Example_NodeNumber/Simple/Run-12         223.0µ ± 87%   294.2µ ± 32%       ~ (p=0.240 n=6)
Example_NodeNumber/Advanced/New-12       76.34m ±  2%   73.95m ±  1%  -3.13% (p=0.002 n=6)
Example_NodeNumber/Advanced/Run-12       46.71µ ±  3%   47.57µ ±  1%       ~ (p=0.065 n=6)
Example_NodeNumber/Simple_Log/New-12                    62.40m ±  1%
Example_NodeNumber/Simple_Log/Run-12                    176.6µ ± 29%
Example_NodeNumber/Advanced_Log/New-12                  73.83m ±  1%
Example_NodeNumber/Advanced_Log/Run-12                  54.63µ ±  1%
geomean                                  2.667m         2.706m        +6.35%
```