@codefromthecrypt 2023-07-28
going to try to add InfoS/ErrorS now. mostly an exercise in copying logic and backfilling tests