@codefromthecrypt 2023-07-29
polished and stubbed structured logging.

next step: implement it