@sanposhiho 2023-07-29
I like this approach. Please go ahead 👍