@codefromthecrypt 2023-07-30
OK I will finish this up, but I won't have time to implement the structured log in the same PR (will leave the signatures for a later change). Will mark ready for review when ready.