@codefromthecrypt 2023-08-02
Added the commit from #69 to proceed further. Will try to finish structured logging in the same PR