@codefromthecrypt 2023-08-02
Ok I've gotten this far, but something's wrong with k/v on the host side. Will look into it later

```
I0802 15:18:53.788082   34631 host.go:219] NodeNumberArgs is successfully applied
I0802 15:18:53.788393   34631 host.go:264] "execute PreScore on NodeNumber plugin" pod="(MISSING)"
I0802 15:18:53.788411   34631 host.go:264] "execute Score on NodeNumber plugin" pod="(MISSING)"
```