@codefromthecrypt 2023-08-02
So, what's left here is..

* debug why KV isn't working on the host (only gets the key not the value)
  * after debugging, in `newNodeNumberPlugin` revert to `LogSeverity: 3, // FatalLog`
* backfill tests for klog host functions (like exist for others, possibly making a tinygo or wat file in testdata)
* See if there's a way to integration test klog in e2e/scheduler (like capture stderr, though might be annoying due to timestamps)
* Add `ExampleInfoS` like we have `ExampleInfo` with use like in NodeNumberPlugin
* Add `ExampleError` and `ExampleErrorS` like we have `ExampleInfo` from some existing code in a real plugin
* once #69 is merged rebase this