@evacchi 2023-08-03
I pushed 4 distinct commits so that you can pick any of these changes if you think they help:

1. fix for the logging issue (just a misconfigured test)
2. capture klog and add the assertions
3. add examples -- but I am not sure I did what you had in mind: the output is empty as it was for the other examples
4. I tried to draft another plugin under `all.go` for tinygo `logPlugin` which is just the same as `params` except it should also log them, but I did not figure out where to put the actual test cases; also it felt a bit redundant because of (2); but maybe you had something different in mind

feel free to drop all of them if they are not useful!