@codefromthecrypt 2023-08-04
Thanks for all the help @evacchi I will go through and do any buttoning up once #69 merges, which might be monday due to some of us having typhoon displacement ;)