@codefromthecrypt 2023-08-05
squashed work so far and will move this to the finish line soon. Thanks a lot @evacchi!