@codefromthecrypt 2023-08-06
@sanposhiho I'm marking ready for review even if I have some test cases to backfill. You can see in the PR desc that logging enabled vs disabled makes a difference in performance, which means our ABI is working! Thanks to @evacchi for the collaboration on this.