@codefromthecrypt 2023-08-07
ok ready. only other refactoring will be separate PR as I want to see if we can avoid using "wat" format too much.

The main thing here is we can see that the advanced plugin perf is within the margin of error (comparable to before), and when logging is switched on, it drops 13% (7us) per run, which makes sense as logging is expensive.